requirements.txt
attrs==17.4.0
bleach==2.1.3
colander==0.9.9
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
Flask==0.10.1
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
iso8601==0.1.8
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lark-parser==0.5.6
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
more-itertools==4.1.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.5.0
python-dateutil==2.7.2
python-intervals==1.3.2
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
redis==2.8.0
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
translationstring==1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.0
